CI debugs

+++ src/sundomeigest/ArnI/sundomeigest_arni.c
@@ -336,7 +336,7 @@ SUNErrCode SUNDomEigEstimate_ArnI(SUNDomEigEstimator DEE, suncomplextype* dom_ei
   }
 
   /* Sort the array using qsort */
-  qsort(ArnI_CONTENT(DEE)->LAPACK_arr, n, sizeof(suncomplextype), domeig_Compare);
+  qsort(ArnI_CONTENT(DEE)->LAPACK_arr, n, sizeof(suncomplextype), (int)domeig_Compare);
 
   /* Update the original arrays */
   for (i = 0; i < n; i++)
